fix: dialog delete

+++ src/app/modules/cadastre/batch/batch.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { PageEvent } from '@angular/material/paginator';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { BatchPage } from 'src/app/model/batch-page';
 import { BatchService } from 'src/app/service/batch.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   templateUrl: './batch.component.html',
@@ -19,7 +21,8 @@ export class BatchComponent implements OnInit {
     private batchService: BatchService,
     private formBuilder: UntypedFormBuilder,
     private router: Router,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    public dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -53,20 +56,28 @@ export class BatchComponent implements OnInit {
   }
 
   onDelete(id: number) {
-    this.batchService.delete(id).subscribe(
-      () => {
-        this.onFilter();
-      },
-      (e) => {
-        if (e?.status === 400 || e?.status === 404) {
-          this.snackBar.open(e?.error?.userMessage, '', { duration: 5000 });
-        } else {
-          this.snackBar.open('Erro ao excluir o lote', '', {
-            duration: 5000,
-          });
-        }
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover o lote?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.batchService.delete(id).subscribe(
+          () => {
+            this.onFilter();
+          },
+          (e) => {
+            if (e?.status === 400 || e?.status === 404) {
+              this.snackBar.open(e?.error?.userMessage, '', { duration: 5000 });
+            } else {
+              this.snackBar.open('Erro ao excluir o lote', '', {
+                duration: 5000,
+              });
+            }
+          }
+        );
       }
-    );
+    });
   }
 
   onPageInfo(pageEvent: PageEvent): void {
